Dynamic hitboxes
Pressing the A button in the air (placeholder ofc) will now create a
hitbox around your body for 5 frames and put you in the AirAttack state.
In this state you have full air control but can't do anything else. Next
step might be creating a hitbox manager so we can have hitboxes that
change over the course of their duration. Hitbox knockback vector also
depends on the direction you're facing, so the hitbox should always
knock the opponent forward.